worker/requirements.txt
names-generator==0.1.0
python-dotenv==0.21.0
requests==2.28.1
urllib3==1.26.13
